moves retrieve user test to UserTest.php
