+++ tests/AuthTest.php
@@ -192,47 +192,4 @@ class AuthTest extends TestCase
         $tokenCount = Token::where('key', '=', $data['data']['session_token'])->count();
         $this->assertEquals($tokenCount, 1);
     }
-
-    /**
-     * Tests GET /users/{term}/{id} endpoint works correctly
-     */
-    public function testRetrieveUser()
-    {
-        // User info
-        $user = User::find('5430e850dt8hbc541c37tt3d');
-
-        // GET /users/_id/<user_id>
-        $response = $this->call('GET', 'v1/users/_id/'.$user->_id, [], [], [], $this->serverMissingToken);
-        $content = $response->getContent();
-        $data = json_decode($content, true);
-
-        // Assert response is 200 and has expected data
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertJson($content);
-        $this->assertArrayHasKey('_id', $data['data'][0]);
-
-        // GET /users/mobile/<mobile>
-        $response = $this->call('GET', 'v1/users/mobile/'.$user->mobile, [], [], [], $this->serverMissingToken);
-        $content = $response->getContent();
-        $data = json_decode($content, true);
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertJson($content);
-        $this->assertArrayHasKey('mobile', $data['data'][0]);
-
-        // GET /users/email/<email>
-        $response = $this->call('GET', 'v1/users/email/'.$user->email, [], [], [], $this->serverMissingToken);
-        $content = $response->getContent();
-        $data = json_decode($content, true);
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertJson($content);
-        $this->assertArrayHasKey('email', $data['data'][0]);
-
-        // GET /users/drupal_id/<drupal_id>
-        $response = $this->call('GET', 'v1/users/drupal_id/'.$user->drupal_id, [], [], [], $this->serverMissingToken);
-        $content = $response->getContent();
-        $data = json_decode($content, true);
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertJson($content);
-        $this->assertArrayHasKey('drupal_id', $data['data'][0]);
-    }
 }
